+++ R/netleague.R
@@ -181,7 +181,7 @@
 #'   studlab = id, data = Linde2015, sm = "OR")
 #' #
 #' net1 <- netmeta(p1, common = FALSE,
-#'                 seq = trts, ref = "Placebo", small = "bad")
+#'                 seq = trts, ref = "Placebo", small = "undesirable")
 #' 
 #' # (2) Early remission
 #' #
@@ -190,7 +190,7 @@
 #'   studlab = id, data = Linde2015, sm = "OR")
 #' #
 #' net2 <- netmeta(p2, common = FALSE,
-#'                 seq = trts, ref = "Placebo", small = "bad")
+#'                 seq = trts, ref = "Placebo", small = "undesirable")
 #' 
 #' options(width = 200)
 #' netleague(net1, digits = 2)