+++ R/netmeta-internal.R
@@ -90,3 +90,19 @@ legendabbr <- function(full, abbr, condition,
   }
   invisible(condition & any(diff))
 }
+
+
+setsv <- function(x) {
+  if (is.null(x))
+    res <- "desirable"
+  else {
+    res <- setchar(x, c("good", "bad"), stop.at.error = FALSE)
+    ##
+    if (!is.null(res))
+      res <- switch(res, good = "desirable", bad = "undesirable")
+    else
+      res <- x
+  }
+  ##
+  setchar(res, c("desirable", "undesirable"))
+}