+++ R/plot.netposet.R
@@ -126,7 +126,7 @@
 #'   studlab = id, data = Linde2015, sm = "OR")
 #' #
 #' net1 <- netmeta(p1, common = FALSE,
-#'   seq = trts, ref = "Placebo", small.values = "bad")
+#'   seq = trts, ref = "Placebo", small.values = "undesirable")
 #' 
 #' # (2) Early remission
 #' #
@@ -135,7 +135,7 @@
 #'   studlab = id, data = Linde2015, sm = "OR")
 #' #
 #' net2 <- netmeta(p2, common = FALSE,
-#'   seq = trts, ref = "Placebo", small.values = "bad")
+#'   seq = trts, ref = "Placebo", small.values = "undesirable")
 #' 
 #' # Partial order of treatment rankings
 #' #
@@ -163,7 +163,7 @@
 #'   studlab = id, data = Linde2015, sm = "OR")
 #' #
 #' net3 <- netmeta(p3, common = FALSE,
-#'   seq = trts, ref = "Placebo", small.values = "good")
+#'   seq = trts, ref = "Placebo", small.values = "desirable")
 #' 
 #' # Partial order of treatment rankings (with three outcomes) 
 #' #